utils/datasets/waterbirds/requirements.txt
numpy
pandas
pillow
tqdm
